service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
openai==1.57.0
python-dotenv==1.0.1
audiofile==1.4.0
pytest==8.3.3
httpx==0.27.2
